css-ui-3/caret-color-009.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Basic User Interface Test: caret-color computed values</title>
<link rel="author" title="Florian Rivoal" href="mailto:florian@rivoal.net">
<link rel="help" href="http://www.w3.org/TR/css3-ui/#caret-color">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#currentcolor-color">
<meta name="flags" content="dom">
<meta name="assert" content="Test checks that auto and currentcolor compute to themselves for caret-color">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
#d1 {
	caret-color: auto;
}
#d2 {
	caret-color: currentcolor;
}
</style>
<body>
  <div id=d1></div>
  <div id=d2></div>
  <div id=log></div>

<script>
test(
  function(){
   var d1 = document.getElementById("d1");
   assert_equals(window.getComputedStyle(d1)["caret-color"], "auto");
  }, "The computed value of auto should be auto");
test(
  function(){
   var d2 = document.getElementById("d2");
   assert_equals(window.getComputedStyle(d1)["caret-color"], "currentcolor");
  }, "The computed value of currentcolor should be currentcolor");
</script>
</body>
</html>
